add relation on character_skills create a dedicated table for unallocated_sp and total_sp fix character_id, corporation_id, alliance_id type according to latest CCP recommendation

+++ src/database/migrations/2017_12_21_124934_create_refresh_tokens_table.php
@@ -16,7 +16,7 @@ class CreateRefreshTokensTable extends Migration
 
         Schema::create('refresh_tokens', function (Blueprint $table) {
 
-            $table->integer('character_id')->primary();
+            $table->bigInteger('character_id')->primary();
 
             $table->mediumText('refresh_token');
             $table->json('scopes');
