+++ src/database/migrations/2018_01_04_201942_create_character_standings_table.php
@@ -18,7 +18,7 @@ class CreateCharacterStandingsTable extends Migration
 
             $table->increments('id');
 
-            $table->integer('character_id');
+            $table->bigInteger('character_id');
             $table->integer('from_id');
             $table->enum('from_type', ['agent', 'npc_corp', 'faction']);
             $table->float('standing');
